Implemented expirey and fixed problems with http4s packages

+++ src/main/scala/com/example/offerservice/Models.scala
@@ -1,11 +1,6 @@
 package com.example.offerservice
 
 object Models {
-    case class Offer(
-        name: String,
-        description: String,
-        price: BigDecimal,
-        secondsToExpiry: Long,
-    )
+    case class Offer(name: String, description: String, price: BigDecimal, secondsToExpiry: BigDecimal)
     case class StoredOffer(record: Offer, id: String, creationTime: Long)
 }
